Each Research Page has unique linked page
Go to the reasearchPages and select the page you want to edit (they are numbered in the same order as they are displayed). Edit any of the pages but please don't change page 0, the template. If you want to create a new page then you can copy and paste the 0 template page.